docs: refactor labeledgraph examples
