+++ labeledgraph/example_labeledgraph_test.go
@@ -1,7 +1,10 @@
 package labeledgraph_test
 
 import (
+	"cmp"
 	"fmt"
+	"github.com/lock14/collections/hashset"
+	"github.com/lock14/collections/heap"
 	"github.com/lock14/collections/labeledgraph"
 	"slices"
 )
@@ -352,4 +355,156 @@ func ExampleLabeledGraph_String() {
 	fmt.Println(g2.String())
 	// Output:
 	// [1 -> 2: A]
+}
+
+type PathNode struct {
+	vertex string
+	dist   int
+}
+
+func ExampleLabeledGraph_shortestPath() {
+	// Create a directed graph with string vertices and int edge weights (distances)
+	g := labeledgraph.New[string, int](labeledgraph.Directed())
+
+	// Add edges and their weights
+	g.AddEdge("A", "B", 4)
+	g.AddEdge("A", "C", 2)
+	g.AddEdge("B", "C", 5)
+	g.AddEdge("B", "D", 10)
+	g.AddEdge("C", "E", 3)
+	g.AddEdge("E", "D", 4)
+	g.AddEdge("D", "F", 11)
+
+	// We want to find the shortest path from "A" to all other nodes.
+	start := "A"
+
+	// Priority queue to select the node with the minimum distance
+	pq := heap.New[PathNode](heap.WithComparator(func(a, b PathNode) int {
+		return cmp.Compare(a.dist, b.dist)
+	}))
+
+	// Map to store the shortest known distance to each node
+	distances := make(map[string]int)
+
+	// Map to store the previous node to reconstruct the path
+	previous := make(map[string]string)
+
+	distances[start] = 0
+	pq.Add(PathNode{vertex: start, dist: 0})
+
+	for !pq.Empty() {
+		current := pq.Remove()
+
+		// If we already found a shorter path to this node, skip processing
+		if current.dist > distances[current.vertex] {
+			continue
+		}
+
+		for neighbor := range g.Successors(current.vertex) {
+			weight, _ := g.Label(current.vertex, neighbor)
+			newDist := current.dist + weight
+
+			// If we haven't visited the neighbor, or found a shorter path
+			if d, visited := distances[neighbor]; !visited || newDist < d {
+				distances[neighbor] = newDist
+				previous[neighbor] = current.vertex
+				pq.Add(PathNode{vertex: neighbor, dist: newDist})
+			}
+		}
+	}
+
+	// Print shortest distances from A
+	for _, v := range []string{"A", "B", "C", "D", "E", "F"} {
+		fmt.Printf("Distance from %s to %s: %d\n", start, v, distances[v])
+	}
+
+	// Print shortest path to F
+	path := []string{"F"}
+	curr := "F"
+	for curr != start {
+		curr = previous[curr]
+		path = append([]string{curr}, path...)
+	}
+	fmt.Printf("Shortest path to F: %v\n", path)
+
+	// Output:
+	// Distance from A to A: 0
+	// Distance from A to B: 4
+	// Distance from A to C: 2
+	// Distance from A to D: 9
+	// Distance from A to E: 5
+	// Distance from A to F: 20
+	// Shortest path to F: [A C E D F]
+}
+
+func ExampleLabeledGraph_minimumSpanningTree() {
+	// Create an undirected graph with string vertices and int edge weights
+	g := labeledgraph.New[string, int]()
+
+	// Add edges for a connected network
+	g.AddEdge("A", "B", 4)
+	g.AddEdge("A", "H", 8)
+	g.AddEdge("B", "C", 8)
+	g.AddEdge("B", "H", 11)
+	g.AddEdge("C", "I", 2)
+	g.AddEdge("C", "F", 4)
+	g.AddEdge("C", "D", 7)
+	g.AddEdge("D", "E", 9)
+	g.AddEdge("D", "F", 14)
+	g.AddEdge("E", "F", 10)
+	g.AddEdge("F", "G", 2)
+	g.AddEdge("G", "I", 6)
+	g.AddEdge("G", "H", 1)
+	g.AddEdge("H", "I", 7)
+
+	type Edge struct {
+		u, v   string
+		weight int
+	}
+
+	// Priority queue to select the minimum weight edge
+	pq := heap.New[Edge](heap.WithComparator(func(a, b Edge) int {
+		return cmp.Compare(a.weight, b.weight)
+	}))
+
+	visited := hashset.New[string]()
+	mstEdges := make([]Edge, 0)
+	totalCost := 0
+
+	// Start from arbitrary vertex "A"
+	start := "A"
+	visited.Add(start)
+
+	// Add all edges from the start vertex to the queue
+	for neighbor := range g.Neighbors(start) {
+		weight, _ := g.Label(start, neighbor)
+		pq.Add(Edge{u: start, v: neighbor, weight: weight})
+	}
+
+	for !pq.Empty() {
+		e := pq.Remove()
+
+		// If the destination is already visited, this edge forms a cycle in the MST
+		if visited.Contains(e.v) {
+			continue
+		}
+
+		// Include edge in MST
+		visited.Add(e.v)
+		mstEdges = append(mstEdges, e)
+		totalCost += e.weight
+
+		// Add new outgoing edges from the newly visited vertex to the queue
+		for neighbor := range g.Neighbors(e.v) {
+			if !visited.Contains(neighbor) {
+				weight, _ := g.Label(e.v, neighbor)
+				pq.Add(Edge{u: e.v, v: neighbor, weight: weight})
+			}
+		}
+	}
+
+	fmt.Printf("Total MST Cost: %d\n", totalCost)
+
+	// Output:
+	// Total MST Cost: 37
 }
